rename legal experience to legal credentails
