+++ src/app/onboarding/legal-fellow-credentials/page.tsx
@@ -1,27 +1,26 @@
 'use client';
 
-import BigDataDropdown from '@/components/BigDataDropdown';
-import { BigBlueButton, BigLinkButton } from '@/components/Buttons';
-import { FormControl, FormField, FormItem, FormLabel } from '@/components/Form';
-import Icon from '@/components/Icon';
-import RadioGroup from '@/components/RadioGroup';
-import TextInput from '@/components/TextInput';
-import { states } from '@/data/citiesAndStates';
-import { CardForm, Flex } from '@/styles/containers';
 import { H1Centered } from '@/styles/text';
-import { formatTruthy } from '@/utils/helpers';
-import { useGuardedOnboarding, useOnboardingNavigation } from '@/utils/hooks';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { useRouter } from 'next/navigation';
-import { useMemo } from 'react';
-import { FormProvider, useForm } from 'react-hook-form';
+import RadioGroup from '@/components/RadioGroup';
 import { z } from 'zod';
+import { FormProvider, useForm } from 'react-hook-form';
+import { zodResolver } from '@hookform/resolvers/zod';
+import { FormControl, FormField, FormItem, FormLabel } from '@/components/Form';
+import { CardForm, Flex } from '@/styles/containers';
+import { BigBlueButton, BigLinkButton } from '@/components/Buttons';
+import { useRouter } from 'next/navigation';
+import { useGuardedOnboarding, useOnboardingNavigation } from '@/utils/hooks';
+import { formatTruthy, getCurrentDate, parseDateAlt } from '@/utils/helpers';
+import DateInput from '@/components/DateInput';
+import { useMemo, useState } from 'react';
+import Icon from '@/components/Icon';
 import * as Styles from '../styles';
 
 // zod schema to automate form validation
 const legalExperienceSchema = z.object({
-  stateBarred: z.string({ required_error: 'Required' }),
-  barNumber: z.string({ required_error: 'Must include attorney bar number' }),
+  expectedBarDate: z
+    .date({ required_error: 'Must include expected barred date' })
+    .min(getCurrentDate(), { message: 'Must select a current or future date' }),
   eoirRegistered: z.boolean({ required_error: 'Must select one option' }),
 });
 
@@ -30,79 +29,69 @@ export default function Page() {
   const { backlinkHref } = useOnboardingNavigation();
   const { push } = useRouter();
 
-  // initialize form with data from onboarding context
+  const [expectedBarDate, setExpectedBarDate] = useState<string>(
+    onboarding.profile.expected_bar_date ?? '',
+  );
+
+  // initialize form with values from onboarding context
   const form = useForm<z.infer<typeof legalExperienceSchema>>({
     resolver: zodResolver(legalExperienceSchema),
     defaultValues: {
-      stateBarred: onboarding.profile.state_barred,
-      barNumber: onboarding.profile.bar_number,
+      expectedBarDate: onboarding.profile.expected_bar_date
+        ? new Date(`${onboarding.profile.expected_bar_date}T00:00`)
+        : undefined,
       eoirRegistered: onboarding.profile.eoir_registered,
     },
   });
 
-  const onSubmit = () => {
+  const onValidSubmit = () => {
     push(`/onboarding/${onboarding.flow[4].url}`);
   };
 
   const formValues = form.watch();
   const isEmpty = useMemo(
     () =>
-      !(formValues.stateBarred && formValues.barNumber) ||
+      formValues.expectedBarDate === undefined ||
       formValues.eoirRegistered === undefined,
     [formValues],
   );
 
   return (
     <FormProvider {...form}>
-      <CardForm onSubmit={form.handleSubmit(onSubmit)}>
+      {/* noValidate to prevent default HTML invalid input pop-up */}
+      <CardForm onSubmit={form.handleSubmit(onValidSubmit)} noValidate>
         <Styles.BackLink href={backlinkHref}>
           <Icon type="leftArrow" />
         </Styles.BackLink>
 
-        <H1Centered>Legal Experience</H1Centered>
+        <H1Centered>Legal Credentials</H1Centered>
 
         <Styles.FormFieldsContainer>
           <FormField
             control={form.control}
-            name="stateBarred"
+            name="expectedBarDate"
             render={({ field, fieldState }) => (
               <FormItem>
-                <FormLabel>Which state are you barred in?</FormLabel>
+                <FormLabel>When are you expected to be barred?</FormLabel>
                 <FormControl>
-                  <BigDataDropdown
-                    options={states}
+                  <DateInput
                     error={fieldState.error?.message}
+                    min={parseDateAlt(getCurrentDate())}
+                    value={expectedBarDate}
+                    setValue={setExpectedBarDate}
                     onChange={newValue => {
-                      field.onChange(newValue);
+                      // turn "" into undefined (cannot be parsed to date)
+                      if (!newValue) {
+                        field.onChange(undefined);
+                        onboarding.updateProfile({
+                          expected_bar_date: undefined,
+                        });
+                        return;
+                      }
+                      field.onChange(new Date(`${newValue}T00:00`));
                       onboarding.updateProfile({
-                        state_barred: newValue ?? undefined,
+                        expected_bar_date: newValue,
                       });
-                    }}
-                    defaultValue={onboarding.profile.state_barred}
-                    placeholder="Start typing to filter states..."
-                  />
-                </FormControl>
-              </FormItem>
-            )}
-          />
-
-          <FormField
-            control={form.control}
-            name="barNumber"
-            render={({ field, fieldState }) => (
-              <FormItem>
-                <FormLabel>What is your attorney bar number?</FormLabel>
-                <FormControl>
-                  <TextInput
-                    errorText={fieldState.error?.message}
-                    placeholder="123456"
-                    type="text"
-                    defaultValue={field.value}
-                    onChange={newValue => {
-                      onboarding.updateProfile({
-                        bar_number: newValue,
-                      });
-                      field.onChange(newValue);
                     }}
                   />
                 </FormControl>
@@ -129,7 +118,7 @@ export default function Page() {
                       undefined,
                     )}
                     options={['Yes', 'No']}
-                    error={fieldState.error?.message}
+                    error={fieldState.error ? undefined : ''}
                     onChange={newValue => {
                       const bool = newValue === 'Yes';
                       onboarding.updateProfile({ eoir_registered: bool });
